+++ CHANGELOG.md
@@ -7,10 +7,10 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## [Unreleased]
 
 ### Added
-- Can check whether byte ranges and byte range sets are satisfiable.
+- Nothing
 
 ### Changed
-- Can create suffix byte ranges of length zero.
+- Nothing
 
 ### Deprecated
 - Nothing
@@ -23,6 +23,14 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Security
 - Nothing
+
+## [0.3.0] - 2018-08-01
+
+### Added
+- Can check whether byte ranges and byte range sets are satisfiable.
+
+### Changed
+- Can create suffix byte ranges of length zero.
 
 ## [0.2.0] - 2018-07-27
 
@@ -40,5 +48,6 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - If-Match HTTP header.
 - If-None-Match HTTP header.
 
-[Unreleased]: https://github.com/Stadly/HttpTest/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/Stadly/HttpTest/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/Stadly/HttpTest/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/Stadly/HttpTest/compare/v0.1.0...v0.2.0